Sync.

See merge request self_driving_bus/itriadv!87

+++ src/sensing/camera_grabber/src/TegraAGrabber.cpp
@@ -34,7 +34,7 @@ TegraAGrabber::~TegraAGrabber()
     delete grabber;
     printf("grabber DELETED OK!\n");
   }
-  for (int i = 0; i < cam_ids_.size(); i++)
+  for (size_t i = 0; i < cam_ids_.size(); i++)
   {
     nppiFree(npp8u_ptrs_[i]);
   }
